requirements.txt
absl-py @ file:///C:/ci/absl-py_1623867344508/work
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
blinker==1.4
dataclasses==0.8
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
future==0.18.2
gast==0.2.2
graphviz==0.17
grpcio==1.41.0
h5py==3.1.0
imageio==2.9.0
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
keras==2.6.0
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
matplotlib==3.3.4
numpy==1.18.5
opencv-contrib-python==4.5.3.56
opencv-python==4.1.1.26
pandas==1.1.5
parso==0.5.2
Pillow==8.4.0
protobuf==3.17.2
pytesseract==0.3.8
tensorflow==2.0.3
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.9.4
tesseract==0.1.3
